tron-mcp-server/requirements.txt
# TRON MCP Server 依赖

# MCP SDK
mcp>=1.0.0

# HTTP 客户端
httpx>=0.24.0

# 环境变量
python-dotenv>=1.0.0

# 开发依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
